requirements.txt
numpy
PyContracts==1.8.5
nose
scipy
six
